implement HTTP_request and fix a bug in HTTP_server
About the bug in HTTP_server:
The initialization code of buf in HTTP_server/lib/Server.c callProperHandler()
is wrong. Should be `memset(buf, 0, MAX_HEADER_SIZE);` or just `char buf[MAX_HEADER_SIZE]= {0};`
The wrong code may cause a bug that when POST <http://localhost:8000/postdemo>
with HTTP_request, the HTTP_request won't get the right response.

The reason is:
There are chances that that the client did't send the data to the server once.
For example, only sent "POST /postdemo HTTP/1.1\r\n" for the time being. And
when calling callProperHandler(), the callstack frame has the same position as last
callProperHandler(), thus the buf has the same position, and may contain the full dirty
data of last HTTP request message. And now in the buf there's full HTTP request message
despite the client has only sent the request line. And then in HTTP_server/src/DemoApplication.c
postDemoHandleFunc() the calculation of receivedBodyLength is wrong (gets a negative number)
and postDemoHandleFunc() starts to mistake the incoming client sent data as body and send it
back to the client.